services/auth_service/requirements.txt

fastapi>=0.100,<1.0
uvicorn[standard]>=0.22,<1.0
passlib>=1.7,<2.0
python-jose>=3.3,<4.0
sqlalchemy>=2.0,<3.0
pydantic>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
